requirements.txt
# A股信号灯 - 核心依赖
# ========================
akshare>=1.12.0             # 东方财富数据源（备用）
efinance>=0.5.5             # 东方财富数据源
pandas>=2.0.0               # 数据分析
numpy>=1.24.0               # 数值计算
requests>=2.31.0            # HTTP请求
python-dotenv>=1.0.0        # 环境变量配置
schedule>=1.2.0            # 定时任务调度

# 可选：AI增强（可暂不安装）
# litellm>=1.0.0
# openai>=1.0.0
